packages/research_dispatcher/requirements.txt
supabase==2.10.0
python-dotenv==1.0.0
reportlab==4.0.9
pyyaml==6.0.2
